Move around i18n code
