+++ data/i18n/seeds/en/messages.json
@@ -1,0 +1,3 @@
+{
+
+}
