+++ data/i18n/seeds/es/messages.json
@@ -1,0 +1,3 @@
+{
+
+}
