+++ lib/suma/tasks/i18n.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "rake/tasklib"
-
 require "suma/tasks"
 
 class Suma::Tasks::I18n < Rake::TaskLib
@@ -11,19 +9,19 @@ class Suma::Tasks::I18n < Rake::TaskLib
       desc "Import static_string_keys into the database."
       task :import_static_string_keys do
         Suma.load_app?
-        Suma::I18n::StaticString.import_all_namespaces
+        Suma::I18n::StaticStringIO.import_all_keys
       end
 
       desc "Import seeds into the database."
       task :import do
         Suma.load_app?
-        Suma::I18n.import_seeds
+        Suma::I18n::StaticStringIO.import_seeds
       end
 
       desc "Export seeds from the database."
       task :export do
         Suma.load_app?
-        Suma::I18n.export_seeds
+        Suma::I18n::StaticStringIO.export_seeds
       end
     end
   end
